+++ bin/.gitignore
@@ -1,0 +1,2 @@
+/Ex3/
+/Task/
