+++ bin/App/.gitignore
@@ -1,0 +1,2 @@
+/mainFraim.class
+/package-info.class
