--- src/Ex3/package-info.java
@@ -1,1 +1,0 @@
-package Ex3;
