+++ src/Task/Main.java
@@ -1,4 +1,4 @@
-package Ex3;
+package Task;
 
 public class Main {
 //	public static void main(String [] args) {
